miniproject2/backend/requirements.txt
Django~=5.1.8
djangorestframework-simplejwt~=5.3.1
djangorestframework~=3.15.1
PyPDF2
python-docx
spacy
celery
redis
django-cors-headers
psycopg2-binary